prepare test structure
